user will not have to verify email first.

+++ src/main/java/edu/sjsu/LPOS/service/UserService.java
@@ -15,5 +15,5 @@ public interface UserService {
  
     void deleteUser(Integer id);
     
-    User findUserById(Integer id);
+    User findUserById(Long id);
 }
